Add Validators

+++ T-API.BLL/Concrete/UserManager.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using T_API.BLL.Abstract;
+using T_API.BLL.Validators.User;
 using T_API.Core.DTO.User;
 using T_API.Core.Exception;
 using T_API.DAL.Abstract;
@@ -26,6 +27,13 @@ namespace T_API.BLL.Concrete
 
             try
             {
+                AddUserValidator validator=new AddUserValidator();
+                var validation = validator.Validate(addUserDto);
+                if (!validation.IsValid)
+                {
+                    throw new ValidationException(validation.Errors.ToString());
+                }
+
                 var mappedData = _mapper.Map<UserEntity>(addUserDto);
                 var insertedId = await _userRepository.AddUser(mappedData);
                 if (insertedId == 0)
@@ -47,6 +55,12 @@ namespace T_API.BLL.Concrete
 
             try
             {
+                DeleteUserValidator validator=new DeleteUserValidator();
+                var result = validator.Validate(deleteUserDto);
+                if (!result.IsValid)
+                {
+                    throw new ValidationException(result.Errors.ToString());
+                }
                 var mappedData = _mapper.Map<UserEntity>(deleteUserDto);
                 await _userRepository.DeleteUser(mappedData);
             }
@@ -102,6 +116,12 @@ namespace T_API.BLL.Concrete
         {
             try
             {
+                UpdateUserValidator validator=new UpdateUserValidator();
+                var result = validator.Validate(updateUserDto);
+                if (!result.IsValid)
+                {
+                    throw new ValidationException(result.Errors.ToString());
+                }
                 var mappedData = _mapper.Map<UserEntity>(updateUserDto);
                 await _userRepository.UpdateUser(mappedData);
